dom/html/test/test_bug57600.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=57600
-->
<head>
  <title>Test for Bug 57600</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=57600">Mozilla Bug 57600</a>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 57600 **/
SimpleTest.waitForExplicitFinish();
var count = 0;
function disp(win) {
  var d = win ? win.document : self.testname.document;
  var str = 'You should see this';
  d.open();
  d.write(str);
  d.close();
  is(d.documentElement.textContent, str, "Unexpected text");
  if (++count == 2) {
    SimpleTest.finish();
  }
}
</script>
</pre>
<p id="display">
  <iframe src="javascript:'<body onload=&quot;this.onerror = parent.onerror; parent.disp(self)&quot;></body>'"> 
  </iframe>
  <iframe name="testname" src="javascript:'<body onload=&quot;this.onerror = parent.onerror; parent.disp()&quot;></body>'"> 
  </iframe>
</p>
</body>
</html>
